workers/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
redis==5.0.0
boto3==1.28.85
openai-whisper==20231117
librosa==0.10.0
numpy==1.26.0
scipy==1.11.4
opencv-python==4.8.1.78
ffmpeg-python==0.2.0
requests==2.31.0
pydantic-settings==2.0.3
httpx==0.25.1
rq==1.14.1
python-multipart==0.0.6
